refactor: created icons.tsx and imported all svgs there
