+++ src/app/components/Upload.tsx
@@ -1,6 +1,6 @@
 'use client'
 import React from "react";
-import Cloud from './incons'
+import {icons} from './icons'
 const UploadComponent: React.FC = () => {
   return (
     <div className="max-w-2xl mx-auto">
@@ -10,7 +10,7 @@ const UploadComponent: React.FC = () => {
           className="flex flex-col items-center justify-center w-full h-64 border-2 border-gray-300 border-dashed rounded-lg cursor-pointer bg-gray-50 dark:hover:bg-bray-800 dark:bg-gray-700 hover:bg-gray-100 dark:border-gray-600 dark:hover:border-gray-500 dark:hover:bg-gray-600"
         >
           <div className="flex flex-col items-center justify-center pt-5 pb-6">
-            <Cloud/>
+            <icons.Cloud/>
             <p className="mb-2 text-sm text-gray-500 dark:text-gray-400">
               <span className="font-semibold">Click to upload</span> or drag and drop
             </p>
